+++ src/lib/services/block/index.ts
@@ -1,4 +1,4 @@
-import { useQuery } from "@tanstack/react-query";
+import { useInfiniteQuery, useQuery } from "@tanstack/react-query";
 import type { UseQueryOptions } from "@tanstack/react-query";
 
 import {
@@ -16,6 +16,7 @@ import {
 
 import { getBlockData, getBlocks } from "./api";
 import { getBlockDataLcd, getLatestBlockLcd } from "./lcd";
+import { getBlockDataSequencer, getBlocksSequencer } from "./sequencer";
 
 export const useBlocks = (
   limit: number,
@@ -89,6 +90,37 @@ export const useLatestBlockLcd = () => {
       retry: false,
       refetchOnWindowFocus: false,
       cacheTime: 0,
+    }
+  );
+};
+
+export const useBlocksSequencer = () => {
+  const endpoint = useLcdEndpoint();
+
+  const { data, ...rest } = useInfiniteQuery(
+    [CELATONE_QUERY_KEYS.BLOCKS_SEQUENCER, endpoint],
+    async ({ pageParam }) => getBlocksSequencer(endpoint, pageParam),
+    {
+      getNextPageParam: (lastPage) => lastPage.pagination.nextKey ?? undefined,
+      refetchOnWindowFocus: false,
+    }
+  );
+
+  return {
+    data: data?.pages.flatMap((page) => page.blocks),
+    ...rest,
+  };
+};
+
+export const useBlockDataSequencer = (height: number) => {
+  const endpoint = useLcdEndpoint();
+
+  return useQuery<BlockData>(
+    [CELATONE_QUERY_KEYS.BLOCK_DATA_SEQUENCER, endpoint, height],
+    async () => getBlockDataSequencer(endpoint, height),
+    {
+      retry: false,
+      refetchOnWindowFocus: false,
     }
   );
 };