+++ src/lib/services/nft/nftService.ts
@@ -2,15 +2,8 @@ import type { UseQueryOptions } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
 
 import {
-  useNftsByAccountExpression,
-  useNftsByAccountExpressionOld,
-  useNftsExpression,
-  useNftsExpressionOld,
-} from "../expression";
-import {
   CELATONE_QUERY_KEYS,
   useCelatoneApp,
-  useCurrentChain,
   useNftConfig,
 } from "lib/app-provider";
 import type { HexAddr, HexAddr32, MutateEvent } from "lib/types";
@@ -35,48 +28,26 @@ import {
   getNftTransactions,
   getNftTransactionsCount,
 } from "./nft";
-import {
-  getNftByNftAddressOld,
-  getNftMutateEventsCountOld,
-  getNftMutateEventsOld,
-  getNftsByAccountOld,
-  getNftsCountByAccountOld,
-  getNftsOld,
-  getNftTransactionsCountOld,
-  getNftTransactionsOld,
-} from "./nftOld";
-
-const INITIATION_CHAIN_ID = "initiation-1";
 
 export const useNfts = (
   collectionAddress: HexAddr32,
-  pageSize: number,
+  limit: number,
   offset: number,
   search = ""
 ) => {
   const { chainConfig } = useCelatoneApp();
-  const {
-    chain: { chain_id: chainId },
-  } = useCurrentChain();
-  const expressionNew = useNftsExpression(collectionAddress, search);
-  const expressionOld = useNftsExpressionOld(collectionAddress, search);
-  const expression =
-    chainId === INITIATION_CHAIN_ID ? expressionNew : expressionOld;
 
   return useQuery<Nft[]>(
     [
       CELATONE_QUERY_KEYS.NFTS,
       chainConfig.indexer,
       collectionAddress,
+      limit,
       offset,
-      pageSize,
       search,
-      expression,
     ],
     async () =>
-      chainId === INITIATION_CHAIN_ID
-        ? getNfts(chainConfig.indexer, pageSize, offset, expression)
-        : getNftsOld(chainConfig.indexer, pageSize, offset, expression),
+      getNfts(chainConfig.indexer, collectionAddress, search, limit, offset),
     {
       retry: 1,
       refetchOnWindowFocus: false,
@@ -89,9 +60,6 @@ export const useNftByNftAddress = (
   nftAddress: HexAddr32
 ) => {
   const { chainConfig } = useCelatoneApp();
-  const {
-    chain: { chain_id: chainId },
-  } = useCurrentChain();
 
   return useQuery<NftByNftAddressResponse>(
     [
@@ -101,13 +69,7 @@ export const useNftByNftAddress = (
       nftAddress,
     ],
     async () =>
-      chainId === INITIATION_CHAIN_ID
-        ? getNftByNftAddress(chainConfig.indexer, collectionAddress, nftAddress)
-        : getNftByNftAddressOld(
-            chainConfig.indexer,
-            collectionAddress,
-            nftAddress
-          ),
+      getNftByNftAddress(chainConfig.indexer, collectionAddress, nftAddress),
     {
       retry: 1,
       refetchOnWindowFocus: false,
@@ -145,10 +107,6 @@ export const useNftTransactions = (
   nftAddress: HexAddr32
 ) => {
   const { chainConfig } = useCelatoneApp();
-  const {
-    chain: { chain_id: chainId },
-  } = useCurrentChain();
-
   return useQuery<NftTransactions[]>(
     [
       CELATONE_QUERY_KEYS.NFT_TRANSACTIONS,
@@ -158,9 +116,7 @@ export const useNftTransactions = (
       offset,
     ],
     async () =>
-      chainId === INITIATION_CHAIN_ID
-        ? getNftTransactions(chainConfig.indexer, nftAddress, offset, limit)
-        : getNftTransactionsOld(chainConfig.indexer, nftAddress, offset, limit),
+      getNftTransactions(chainConfig.indexer, nftAddress, limit, offset),
     {
       retry: 1,
       refetchOnWindowFocus: false,
@@ -170,20 +126,13 @@ export const useNftTransactions = (
 
 export const useNftTransactionsCount = (nftAddress: HexAddr32) => {
   const { chainConfig } = useCelatoneApp();
-  const {
-    chain: { chain_id: chainId },
-  } = useCurrentChain();
-
   return useQuery<number>(
     [
       CELATONE_QUERY_KEYS.NFT_TRANSACTIONS_COUNT,
       chainConfig.indexer,
       nftAddress,
     ],
-    async () =>
-      chainId === INITIATION_CHAIN_ID
-        ? getNftTransactionsCount(chainConfig.indexer, nftAddress)
-        : getNftTransactionsCountOld(chainConfig.indexer, nftAddress),
+    async () => getNftTransactionsCount(chainConfig.indexer, nftAddress),
     {
       retry: 1,
       refetchOnWindowFocus: false,
@@ -197,10 +146,6 @@ export const useNftMutateEvents = (
   nftAddress: HexAddr32
 ) => {
   const { chainConfig } = useCelatoneApp();
-  const {
-    chain: { chain_id: chainId },
-  } = useCurrentChain();
-
   return useQuery<MutateEvent[]>(
     [
       CELATONE_QUERY_KEYS.NFT_MUTATE_EVENTS,
@@ -210,9 +155,7 @@ export const useNftMutateEvents = (
       offset,
     ],
     async () =>
-      chainId === INITIATION_CHAIN_ID
-        ? getNftMutateEvents(chainConfig.indexer, nftAddress, offset, limit)
-        : getNftMutateEventsOld(chainConfig.indexer, nftAddress, offset, limit),
+      getNftMutateEvents(chainConfig.indexer, nftAddress, offset, limit),
     {
       retry: 1,
       refetchOnWindowFocus: false,
@@ -222,20 +165,13 @@ export const useNftMutateEvents = (
 
 export const useNftMutateEventsCount = (nftAddress: HexAddr32) => {
   const { chainConfig } = useCelatoneApp();
-  const {
-    chain: { chain_id: chainId },
-  } = useCurrentChain();
-
   return useQuery<number>(
     [
       CELATONE_QUERY_KEYS.NFT_MUTATE_EVENTS_COUNT,
       chainConfig.indexer,
       nftAddress,
     ],
-    async () =>
-      chainId === INITIATION_CHAIN_ID
-        ? getNftMutateEventsCount(chainConfig.indexer, nftAddress)
-        : getNftMutateEventsCountOld(chainConfig.indexer, nftAddress),
+    async () => getNftMutateEventsCount(chainConfig.indexer, nftAddress),
     {
       retry: 1,
       refetchOnWindowFocus: false,
@@ -246,20 +182,13 @@ export const useNftMutateEventsCount = (nftAddress: HexAddr32) => {
 export const useNftsCountByAccount = (accountAddress: HexAddr) => {
   const { chainConfig } = useCelatoneApp();
   const { enabled } = useNftConfig({ shouldRedirect: false });
-  const {
-    chain: { chain_id: chainId },
-  } = useCurrentChain();
-
   return useQuery<number>(
     [
       CELATONE_QUERY_KEYS.NFTS_COUNT_BY_ACCOUNT,
       chainConfig.indexer,
       accountAddress,
     ],
-    async () =>
-      chainId === INITIATION_CHAIN_ID
-        ? getNftsCountByAccount(chainConfig.indexer, accountAddress)
-        : getNftsCountByAccountOld(chainConfig.indexer, accountAddress),
+    async () => getNftsCountByAccount(chainConfig.indexer, accountAddress),
     {
       retry: 1,
       refetchOnWindowFocus: false,
@@ -270,48 +199,32 @@ export const useNftsCountByAccount = (accountAddress: HexAddr) => {
 
 export const useNftsByAccountByCollection = (
   accountAddress: HexAddr,
-  pageSize: number,
+  limit: number,
   offset: number,
   search = "",
   collectionAddress?: HexAddr32,
   options: Pick<UseQueryOptions<NftsByAccountResponse>, "onSuccess"> = {}
 ) => {
   const { chainConfig } = useCelatoneApp();
-  const {
-    chain: { chain_id: chainId },
-  } = useCurrentChain();
-  const expressionNew = useNftsByAccountExpression(
-    accountAddress,
-    collectionAddress,
-    search
-  );
-  const expressionOld = useNftsByAccountExpressionOld(
-    accountAddress,
-    collectionAddress,
-    search
-  );
-  const expression =
-    chainId === INITIATION_CHAIN_ID ? expressionNew : expressionOld;
-
   return useQuery<NftsByAccountResponse>(
     [
       CELATONE_QUERY_KEYS.NFTS_BY_ACCOUNT_BY_COLLECTION,
       chainConfig.indexer,
       accountAddress,
-      pageSize,
+      limit,
       offset,
       collectionAddress,
-      expression,
+      search,
     ],
     async () =>
-      chainId === INITIATION_CHAIN_ID
-        ? getNftsByAccount(chainConfig.indexer, pageSize, offset, expression)
-        : getNftsByAccountOld(
-            chainConfig.indexer,
-            pageSize,
-            offset,
-            expression
-          ),
+      getNftsByAccount(
+        chainConfig.indexer,
+        accountAddress,
+        limit,
+        offset,
+        collectionAddress,
+        search
+      ),
     {
       retry: 1,
       refetchOnWindowFocus: false,